python/requirements.txt
Flask
SpeechRecognition
google-generativeai